Refactor oidc_provider into online_oidc_provider
