+++ apps/authentication/api/tests/group_tests.py
@@ -7,7 +7,7 @@ from apps.authentication.constants import RoleType
 from apps.authentication.models import GroupMember, GroupRole, OnlineGroup
 from apps.authentication.models import OnlineUser as User
 from apps.events.tests.utils import generate_user
-from apps.oidc_provider.test import OIDCTestCase
+from apps.online_oidc_provider.test import OIDCTestCase
 
 
 class GroupRoleTestCase(OIDCTestCase):
