+++ apps/authentication/api/tests/user_data_tests.py
@@ -8,7 +8,7 @@ from apps.authentication.models import Email
 from apps.authentication.models import OnlineUser as User
 from apps.authentication.models import Position, SpecialPosition
 from apps.events.tests.utils import generate_user
-from apps.oidc_provider.test import OIDCTestCase
+from apps.online_oidc_provider.test import OIDCTestCase
 
 
 class EmailTestCase(OIDCTestCase):
