+++ onlineweb4/settings/base.py
@@ -108,8 +108,8 @@ RECAPTCHA_USE_SSL = config("OW4_DJANGO_RECAPTCHA_USE_SSL", cast=bool, default=Tr
 
 
 # oidc_provider - OpenID Connect Provider
-OIDC_USERINFO = 'apps.oidc_provider.claims.userinfo'
-OIDC_EXTRA_SCOPE_CLAIMS = 'apps.oidc_provider.claims.Onlineweb4ScopeClaims'
+OIDC_USERINFO = 'apps.online_oidc_provider.claims.userinfo'
+OIDC_EXTRA_SCOPE_CLAIMS = 'apps.online_oidc_provider.claims.Onlineweb4ScopeClaims'
 
 USE_X_FORWARDED_HOST = config("OW4_DJANGO_DEVELOPMENT_HTTPS", cast=bool, default=False)
 SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https') if USE_X_FORWARDED_HOST else None
