+++ onlineweb4/settings/rest_framework.py
@@ -2,7 +2,7 @@
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
         'rest_framework.authentication.SessionAuthentication', # Allows users to be logged in to browsable API
-        'apps.oidc_provider.authentication.OidcOauth2Auth', # Allows user to be logged in with open-id
+        'apps.online_oidc_provider.authentication.OidcOauth2Auth', # Allows user to be logged in with open-id
     ),
     'DEFAULT_FILTER_BACKENDS': (
         'django_filters.rest_framework.DjangoFilterBackend',
